requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
Django==4.0.6
psycopg2-binary==2.9.3
sqlparse==0.4.2
django-dynamic-db-router==0.3.0
psycopg2-binary==2.9.3
postgis==1.0.4
geos==0.2.3
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-gis==1.0
django-leaflet==0.28.3


